Changes to support local testing

+++ src/main.rs
@@ -8,10 +8,10 @@ use tokio::time::Duration;
 use walkdir::WalkDir;
 
 mod bundle_struct;
-// FIXME mod deploy;
-// FIXME mod grafana;
+mod deploy;
+mod grafana;
 mod hdx;
-// FIXME mod headless_browser;
+mod headless_browser;
 mod output_struct;
 mod validate;
 
@@ -25,6 +25,25 @@ lazy_static! {
         .expect("BUNDLE_TESTING_USERNAME environment variable must be set");
     static ref BUNDLE_TESTING_PASSWORD: String = std::env::var("BUNDLE_TESTING_PASSWORD")
         .expect("BUNDLE_TESTING_PASSWORD environment variable must be set");
+    static ref IS_LOCAL: bool = {
+        let args: Vec<String> = std::env::args().collect();
+        args.contains(&"--local".to_string())
+    };
+    static ref FOR_MARKETPLACE: bool = {
+        let args: Vec<String> = std::env::args().collect();
+        args.contains(&"--marketplace".to_string())
+    };
+    static ref MATCH_ONLY: String = {
+        let mut value = "".to_string();
+        let args: Vec<String> = std::env::args().collect();
+        for i in 1..args.len() {
+            if !args[i].starts_with("--") {
+                value = args[i].to_string();
+                break;
+            }
+        }
+        value.to_string()
+    };
 }
 
 //pub const GRAFANA_LOCATION: &str = "host.docker.internal:3000";
@@ -33,12 +52,8 @@ pub const GRAFANA_LOCATION: &str = "localhost:3000";
 
 #[tokio::main]
 async fn main() {
-    println!("Hello -- Cluster is {}", *BUNDLE_TESTING_CLUSTER);
-
-    // We only check bundles at the root directory
     let bundle_list = find_bundle_files();
 
-    println!("list={:?}", bundle_list);
     for b in &bundle_list {
         let path = PathBuf::from(b);
         let file_path = path
@@ -54,13 +69,15 @@ async fn main() {
             }
         };
 
-        // TESTING
-        //if !bundle.name.contains("zuplo") {
-        //    continue;
-        //}
+        if !MATCH_ONLY.is_empty() {
+            if !bundle.name.contains(&*MATCH_ONLY) {
+                println!("Ignoring {}", bundle.name);
+                continue;
+            }
+        }
 
         let base_dir = file_path.replace("./", "").replace("/bundle.json", "");
-        println!("base_dir={base_dir}");
+        println!("Testing {}", bundle.name);
 
         match validate_bundle(&base_dir, &bundle).await {
             Ok(_) => (),
@@ -78,15 +95,15 @@ async fn main() {
 
 // These are all of our tests...
 async fn validate_bundle(base: &str, bundle: &Bundle) -> Result<(), String> {
-    /* FIXME
-    match grafana::container::kill().await {
-        Ok(_) => (),
-        Err(e) => {
-            eprintln!("Failed to kill the Grafana container... error={e}");
-            std::process::exit(1);
+    if *IS_LOCAL {
+        match grafana::container::kill().await {
+            Ok(_) => (),
+            Err(e) => {
+                eprintln!("Failed to kill the Grafana container... error={e}");
+                std::process::exit(1);
+            }
         }
     }
-    */
 
     match validate::no_duplicate_tokens::run(bundle).await {
         Ok(_) => (),
@@ -108,7 +125,12 @@ async fn validate_bundle(base: &str, bundle: &Bundle) -> Result<(), String> {
         Err(e) => return Err(format!("Found bad checksum: error={e}")),
     }
 
-    /*
+    match validate::sample_data_exists::run(base, bundle).await {
+        Ok(_) => (),
+        Err(e) => return Err(format!("No sample data: error={e}")),
+    }
+
+    if *IS_LOCAL {
         match grafana::container::start().await {
             Ok(_) => (),
             Err(e) => {
@@ -148,7 +170,7 @@ async fn validate_bundle(base: &str, bundle: &Bundle) -> Result<(), String> {
                 "Dashboard Errors={datasource_error_count} NoDataErrors={nodata_error_count}"
             ));
         }
-    */
+    }
 
     println!("SUCCESS");
